refactor: renames system.py to system_status.py and splits control functionality
Renames system.py to system_status.py to better reflect its purpose and match
the SystemStatus class name it defines. This is part of a larger refactor that
separates system status (state management) from system control (actions).

Changes:
- Renames system.py to system_status.py using git mv
- Moves all control functionality (start_sprinkler, stop_system) to SystemController
- Updates imports in all router files to use system_status
- Fixes scheduler route tests to properly mock SystemStatus properties
- Updates sprinkler route tests to target system_controller for control operations
- Verifies all route tests pass (28 tests)

This change improves separation of concerns:
- SystemStatus: handles state management and configuration
- SystemController: handles hardware control and system operations

Note: test_system_services.py will be updated in a separate commit to handle
this architectural separation.
