+++ tests/test_sprinklers_route.py
@@ -3,6 +3,7 @@ from typing import List
 
 from context import isprinklr
 from isprinklr.schemas import SprinklerCommand, SprinklerConfig
+from isprinklr.system_status import SystemStatus
 from isprinklr.main import app
 
 client = TestClient(app)
@@ -58,14 +59,14 @@ def test_update_sprinklers_server_error(mocker):
     }
 
 def test_start_sprinkler(mocker):
-    mocker.patch('isprinklr.routers.sprinklers.system_status.start_sprinkler', return_value=True)
+    mocker.patch('isprinklr.routers.sprinklers.system_controller.start_sprinkler', return_value=True)
     sprinkler: SprinklerCommand = {"zone": 1, "duration": 300}  # 5 minutes in seconds
     response = client.post("/api/sprinklers/start", json=sprinkler)
     assert response.status_code == 200
     assert response.json() == {"message": "Zone 1 started"}
 
 def test_start_sprinkler_while_running(mocker):
-    mocker.patch('isprinklr.routers.sprinklers.system_status.start_sprinkler', 
+    mocker.patch('isprinklr.routers.sprinklers.system_controller.start_sprinkler', 
                 side_effect=Exception("Failed to start sprinkler, system already active. Active zone: 2"))
     sprinkler: SprinklerCommand = {"zone": 1, "duration": 60}  # 1 minute in seconds
     response = client.post("/api/sprinklers/start", json=sprinkler)
@@ -73,13 +74,13 @@ def test_start_sprinkler_while_running(mocker):
     assert response.json() == {"detail": "Failed to start sprinkler, see logs for details"}
 
 def test_stop_system(mocker):
-    mocker.patch('isprinklr.routers.sprinklers.system_status.stop_system', return_value=True)
+    mocker.patch('isprinklr.routers.sprinklers.system_controller.stop_system', return_value=True)
     response = client.post("/api/sprinklers/stop")
     assert response.status_code == 200
     assert response.json() == {"message": "System stopped"}
 
 def test_stop_system_error(mocker):
-    mocker.patch('isprinklr.routers.sprinklers.system_status.stop_system', 
+    mocker.patch('isprinklr.routers.sprinklers.system_controller.stop_system', 
                 side_effect=Exception("Failed to stop system"))
     response = client.post("/api/sprinklers/stop")
     assert response.status_code == 500
